Create xunit.runner.json
Configuring xUnit.net with JSON: 

https://xunit.github.io/docs/configuring-with-json.html

Adding the configuration file:
To use JSON-based configuration, you must instruct Visual Studio to copy the JSON file into the output folder of your project.

Add a new JSON file to root of your test project. Name the file xunit.runner.json. (If your version of Visual Studio does not have "JSON file" as an option for a new file, just choose to add a text file, and make sure it has the right name.)

Right click on the newly added file in Solution Explorer, and choose Properties. 

In the Properties window, change the setting for Copy to Output Directory to Copy if newer. This ensures that the file to copied into your bin folder where the xUnit.net runners can find it.